Move config image opacity cast to setWatermarkImageOpacity method (#3727)

+++ app/code/core/Mage/Catalog/Helper/Image.php
@@ -143,7 +143,7 @@ class Mage_Catalog_Helper_Image extends Mage_Core_Helper_Abstract
             Mage::getStoreConfig("design/watermark/{$this->_getModel()->getDestinationSubdir()}_image")
         );
         $this->setWatermarkImageOpacity(
-            (int) Mage::getStoreConfig("design/watermark/{$this->_getModel()->getDestinationSubdir()}_imageOpacity")
+            Mage::getStoreConfig("design/watermark/{$this->_getModel()->getDestinationSubdir()}_imageOpacity")
         );
         $this->setWatermarkPosition(
             Mage::getStoreConfig("design/watermark/{$this->_getModel()->getDestinationSubdir()}_position")
@@ -488,12 +488,12 @@ class Mage_Catalog_Helper_Image extends Mage_Core_Helper_Abstract
     /**
      * Set watermark image opacity
      *
-     * @param int $imageOpacity
+     * @param int|string $imageOpacity
      * @return $this
      */
     public function setWatermarkImageOpacity($imageOpacity)
     {
-        $this->_watermarkImageOpacity = $imageOpacity;
+        $this->_watermarkImageOpacity = (int) $imageOpacity;
         $this->_getModel()->setWatermarkImageOpacity($imageOpacity);
         return $this;
     }
